Uml consistency

See merge request COMP2511/21T3/groups/F11A_AVOCADO/project!143

+++ src/main/java/dungeonmania/model/entities/DefenceEquipment.java
@@ -18,7 +18,7 @@ public abstract class DefenceEquipment extends Equipment {
         this.defenceMultiplier = defenceMultiplier;
     }
 
-    public double getBonusAttackDamage() {
+    public int getBonusAttackDamage() {
         return bonusAttackDamage;
     }
